+++ lib/src/model/user.dart
@@ -202,10 +202,10 @@ class UserLinks extends ModelBase {
   final Uri self;
   final Uri html;
   final Uri photos;
-  final Uri likes;
+  final Uri? likes;
   final Uri portfolio;
-  final Uri followers;
-  final Uri following;
+  final Uri? followers;
+  final Uri? following;
 
   @override
   Map<String, dynamic> toJson() {
@@ -220,28 +220,16 @@ class UserLinks extends ModelBase {
     };
   }
 
-  factory UserLinks.fromJson(Map<String, dynamic>? json) {
-    if (json == null) {
-      return UserLinks(
-        self: Uri(),
-        html: Uri(),
-        photos: Uri(),
-        likes: Uri(),
-        portfolio: Uri(),
-        followers: Uri(),
-        following: Uri(),
-      );
-    }
-
+  factory UserLinks.fromJson(Map<String, dynamic> json) {
     return UserLinks(
       source: json,
-      self: (json['self'] as String?)?.let(Uri.parse) ?? Uri(),
-      html: (json['html'] as String?)?.let(Uri.parse) ?? Uri(),
-      photos: (json['photos'] as String?)?.let(Uri.parse) ?? Uri(),
-      likes: (json['likes'] as String?)?.let(Uri.parse) ?? Uri(),
-      portfolio: (json['portfolio'] as String?)?.let(Uri.parse) ?? Uri(),
-      followers: (json['followers'] as String?)?.let(Uri.parse) ?? Uri(),
-      following: (json['following'] as String?)?.let(Uri.parse) ?? Uri(),
+      self: (json['self'] as String).let(Uri.parse),
+      html: (json['html'] as String).let(Uri.parse),
+      photos: (json['photos'] as String).let(Uri.parse),
+      likes: (json['likes'] as String?)?.let(Uri.parse),
+      portfolio: (json['portfolio'] as String).let(Uri.parse),
+      followers: (json['followers'] as String?)?.let(Uri.parse),
+      following: (json['following'] as String?)?.let(Uri.parse),
     );
   }
 }