+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.0
+
+- breaking change: make `likes`, `followers` and `following` of `UserLinks`
+  nullable to adapt to changed API response (#23)
+
 ## 2.2.0
 
 - feat: add `Photos.download(location)` (#22)